refactor(matnode): Move material designators to private header
